app.py
import streamlit as st
import pandas as pd
from PIL import Image
import hashlib
import yaml
from yaml.loader import SafeLoader
import streamlit_authenticator as stauth
import io
import base64

# Embedded Excel data (from your provided sheets)
excel_data = {
    'Loupes Only': pd.DataFrame({
        'Telescope': ['RDH Elite', 'RDH Elite', 'RDH Ergo 3.0', ...],  # Full data from SHEET id=2
        # (I've truncated for brevity; insert full data from your <DOCUMENT> here in actual code)
    }),
    # Similarly for other sheets: 'Omni Optic', 'Light Systems', 'Accessories', 'School Bundles'
    # Use pd.read_excel(io.BytesIO(base64.b64decode('your_base64_excel_here'))) if preferring file embed
}

# Logo (upload your PNG to repo; this is placeholder display)
st.image("orascoptic_logo.png", width=250)  # Assume uploaded to repo

# Password Authentication Setup
names = ["User"]
usernames = ["user"]
passwords = ["pass123"]  # Change this!
hashed_passwords = stauth.Hasher(passwords).generate()

authenticator = stauth.Authenticate(
    dict(zip(usernames, zipped(hashed_passwords, names))),
    "cookie_name",
    "signature_key",
    cookie_expiry_days=30
)

name, authentication_status, username = authenticator.login("Login", "main")

if authentication_status:
    # The rest of your app code here (from previous app.py)
    st.title("Orascoptic Price Search")
    # ... (insert full app logic: modes, dropdowns, add_to_list, etc.)
    authenticator.logout("Logout", "sidebar")

elif authentication_status is False:
    st.error("Username/password is incorrect")
elif authentication_status is None:
    st.warning("Please enter your username and password")