requirements.txt
streamlit
pandas
openpyxl
pillow
PyYAML
streamlit-authenticator 